only float values are now stored in Items
